nlp-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
httpx==0.25.1
spacy==3.7.2
sentence-transformers==2.2.2
torch==2.1.1
faiss-cpu==1.7.4 